Refactor Webpack config
