+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "assets": "sh scripts/assets.sh",
+    "build": "yarn webpack --config ./webpack/webpack.config.js",
     "deploy-staging": "sh scripts/deploy-staging.sh",
     "deploy-storybook": "storybook-to-ghpages",
     "deploy": "sh scripts/deploy.sh",
@@ -41,7 +42,6 @@
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
       "prettier --write",
-      "eslint --fix",
       "git add"
     ]
   },
@@ -226,6 +226,7 @@
     "webpack-command": "^0.4.1",
     "webpack-dev-middleware": "^3.6.2",
     "webpack-hot-middleware": "^2.24.3",
+    "webpack-merge": "^4.2.1",
     "webpack-notifier": "^1.5.1"
   },
   "optionalDependencies": {
