+++ scripts/assets.sh
@@ -4,7 +4,7 @@ set -e -x
 
 rm -rf public/assets
 mkdir public/assets
-NODE_ENV=production webpack
+NODE_ENV=production yarn build
 stylus \
   $(find src/assets -name '*.styl') \
   --compress \
