--- webpack.helpers.js
@@ -1,60 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-
-const { NODE_ENV } = process.env;
-const isDevelopment = NODE_ENV === 'development';
-
-function findAssets(basePath) {
-  const files = fs.readdirSync(path.join(process.cwd(), basePath));
-
-  // Filter out .styl files
-  const validAssets = (file) => {
-    const whitelist = [
-      '.js',
-      '.coffee',
-    ];
-
-    const isValid = whitelist.some(extension =>
-      extension === path.extname(file));
-
-    return isValid;
-  };
-
-  /**
-   * Construct key/value pairs representing Webpack entrypoints; e.g.,
-   * { desktop: [ path/to/desktop.js ] }
-   */
-  return files
-    .filter(validAssets)
-    .reduce((assetMap, file) => {
-      const fileName = path.basename(file, path.extname(file));
-      const asset = {
-        [fileName]: [
-          path.join(__dirname, basePath, file),
-        ],
-      };
-
-      // Load oldschool global module dependencies
-      asset[fileName].unshift('./src/lib/global_modules');
-
-      if (isDevelopment) {
-        asset[fileName].unshift('webpack-hot-middleware/client?reload=true');
-      }
-
-      return {
-        ...assetMap,
-        ...asset,
-      };
-    }, {});
-}
-
-function getEntrypoints() {
-  return {
-    ...findAssets('./src/assets'),
-  };
-}
-
-module.exports = {
-  getEntrypoints,
-  findAssets,
-};
